requirements-test.txt
pytest
pytest-httpx
pytest-asyncio
pytest-homeassistant-custom-component>=0.13